socket-io/requirements-socket.txt
psycopg2==2.9.6
aiohttp==3.8.4
python-socketio==5.8.0
requests==2.31.0
python-dotenv==1.0.0
vinorm==2.0.7
pymongo==4.3.3
unidecode==1.3.6
```